Move most of readline/multiline related code into cli_readline.go
